Move Wallet module to top level
