+++ cooked-validators/src/Cooked/Attack/DoubleSat.hs
@@ -9,9 +9,9 @@ module Cooked.Attack.DoubleSat where
 import Control.Monad
 import Cooked.Attack.Tweak
 import Cooked.MockChain.Monad
-import Cooked.MockChain.Wallet
 import Cooked.Output
 import Cooked.Skeleton
+import Cooked.Wallet
 import Data.List
 import Data.Map (Map)
 import qualified Data.Map as Map
