+++ cooked-validators/src/Cooked/MockChain/Monad/Direct.hs
@@ -24,9 +24,9 @@ import Control.Monad.State.Strict
 import Cooked.MockChain.Monad
 import Cooked.MockChain.Monad.GenerateTx
 import Cooked.MockChain.UtxoState
-import Cooked.MockChain.Wallet
 import Cooked.Output
 import Cooked.Skeleton
+import Cooked.Wallet
 import Data.Default
 import Data.Function (on)
 import Data.List
@@ -224,7 +224,7 @@ runMockChainTFrom i0 =
   fmap (fmap $ second mcstToUtxoState) . runMockChainTRaw def (mockChainSt0From i0)
 
 -- | Executes a 'MockChainT' from the canonical initial state and environment. The canonical
---  environment uses the default 'SlotConfig' and @[Cooked.MockChain.Wallet.wallet 1]@ as the sole
+--  environment uses the default 'SlotConfig' and @[Cooked.Wallet.wallet 1]@ as the sole
 --  wallet signing transactions.
 runMockChainT :: (Monad m) => MockChainT m a -> m (Either MockChainError (a, UtxoState))
 runMockChainT = runMockChainTFrom def
