+++ cooked-validators/src/Cooked/MockChain/Testing.hs
@@ -10,7 +10,7 @@ import Control.Monad
 import Cooked.MockChain.Monad.Direct
 import Cooked.MockChain.Monad.Staged
 import Cooked.MockChain.UtxoState
-import Cooked.MockChain.Wallet
+import Cooked.Wallet
 import Data.Default
 import qualified Data.Text as T
 import Debug.Trace
