+++ cooked-validators/src/Cooked/Pretty.hs
@@ -68,9 +68,9 @@ module Cooked.Pretty where
 import Control.Arrow (second)
 import Cooked.Currencies (permanentCurrencySymbol, quickCurrencySymbol)
 import Cooked.MockChain.UtxoState
-import Cooked.MockChain.Wallet
 import Cooked.Output
 import Cooked.Skeleton
+import Cooked.Wallet
 import Data.Default
 import Data.Function (on)
 import qualified Data.List as List
