+++ cooked-validators/tests/Cooked/Attack/Tweak/CommonSpec.hs
@@ -4,6 +4,7 @@ import Cooked.Attack.Tweak.Common
 import Cooked.MockChain
 import Cooked.MockChain.Testing
 import Cooked.Skeleton
+import Cooked.Wallet
 import Data.Default
 import qualified Ledger.Ada as L
 import qualified Ledger.Value as L
