+++ cooked-validators/tests/Spec.hs
@@ -4,7 +4,7 @@ import qualified Cooked.InlineDatumsSpec as InlineDatumsSpec
 import qualified Cooked.LtlSpec as LtlSpec
 -- import qualified Cooked.MockChain.Monad.StagedSpec as StagedSpec
 -- import qualified Cooked.MockChain.UtxoStateSpec as UtxoStateSpec
--- import qualified Cooked.MockChain.WalletSpec as WalletSpec
+-- import qualified Cooked.WalletSpec as WalletSpec
 -- import qualified Cooked.OutputReorderingSpec as OutputReorderingSpec
 -- import qualified Cooked.QuickValueSpec as QuickValueSpec
 import Test.Tasty
